removed getTimeZone function, folded logic into pelican.go DiscoverPelicans method
